+++ apps/web/hooks/useFeatureFlags.ts
@@ -59,3 +59,12 @@ export function useOnboardingVariant() {
     "control"
   );
 }
+
+export type PricingCopyVariant = "control" | "free-trial-emphasis";
+
+export function usePricingCopyVariant() {
+  return (
+    (useFeatureFlagVariantKey("pricing-copy") as PricingCopyVariant) ||
+    "control"
+  );
+}
